Reorganize core app modules into app_core package
