+++ STRUCTURE.md
@@ -1,29 +1,34 @@
 # Project Structure (Schema Overview)
 
-The SyMo system tray application is organized as a set of cooperating modules. The diagram below highlights the main
-packages and their roles in the runtime flow.
+The SyMo repository is now grouped so core application logic lives under one folder (`app_core/`),
+while documentation and external notification adapters remain in dedicated top-level directories.
 
 ```
 SyMo/
+├─ app_core/               # core application package
+│  ├─ __init__.py
+│  ├─ app.py               # SystemTrayApp runtime, menu wiring, periodic updates
+│  ├─ dialogs.py           # GTK settings dialog; log export, reset counters, notification settings
+│  ├─ power_control.py     # power actions + confirmation/scheduling dialogs
+│  ├─ system_usage.py      # psutil-based system metrics (temp, CPU, RAM/SWAP, disk, network, uptime)
+│  ├─ click_tracker.py     # thread-safe keyboard/mouse counters
+│  ├─ constants.py         # app constants, paths, defaults, language list
+│  ├─ localization.py      # translation helpers and language selection
+│  ├─ language.py          # localized string dictionaries
+│  └─ logging_utils.py     # log rotation utilities
+│
 ├─ notifications/
-│  ├─ __init__.py          # convenience exports for notifiers
-│  ├─ telegram.py          # bot config, polling, commands, periodic status with metrics + click counts
-│  └─ discord.py           # webhook config, periodic/test messages, shared persistence with UI dialog
+│  ├─ __init__.py          # convenience exports
+│  ├─ telegram.py          # Telegram notifier config, polling, command handling
+│  └─ discord.py           # Discord webhook notifier config + send logic
 │
-├─ dialogs.py              # GTK settings dialog; export logs, reset counters, manage visibility/language/notifications
-├─ power_control.py        # PowerControl actions and confirmation/scheduling dialogs
-├─ system_usage.py         # psutil-based system metrics (temperature, CPU, RAM, SWAP, Disk, network, uptime)
-├─ click_tracker.py        # thread-safe counters with increment/reset helpers for input events
-├─ constants.py            # app names, supported languages, polling intervals, filesystem paths
-├─ localization.py         # language selection utilities and translation wrapper
-├─ language.py             # localized string dictionaries referenced by UI + notifications
+├─ docs/                   # project documentation
 │
-├─ tests/                  # pytest suites (click tracking, localization consistency, Discord notifier)
-│
-├─ build.sh                # package/build helper
-├─ uninstall-symo.sh       # cleanup helper
-├─ requirements.txt        # Python dependencies (GTK/AppIndicator, psutil, requests, pynput, etc.)
+├─ tests/                  # pytest suites
+├─ app.py                  # thin launcher for app_core.app
+├─ build.sh                # build helper
+├─ uninstall-symo.sh       # uninstall helper
+├─ requirements.txt        # Python dependencies
 ├─ logo.png, img.png       # branding and README visuals
 └─ README.md               # quickstart and usage overview
 ```
-
