+++ tests/test_language_keys.py
@@ -1,5 +1,5 @@
-from constants import SUPPORTED_LANGS
-from language import LANGUAGES
+from app_core.constants import SUPPORTED_LANGS
+from app_core.language import LANGUAGES
 
 
 def test_supported_languages_match_translations():
